WIP: More SimpleTest fixes
